refactor: make addMaxTablesInDashboardPageHeight be module
